chore: rename SidebarDetail as MailListItem
