+++ frontend/components.d.ts
@@ -36,6 +36,7 @@ declare module 'vue' {
     MailAccountModal: typeof import('./src/components/Modals/MailAccountModal.vue')['default']
     MailDate: typeof import('./src/components/MailDate.vue')['default']
     MailDetailsPopover: typeof import('./src/components/MailDetailsPopover.vue')['default']
+    MailListItem: typeof import('./src/components/MailListItem.vue')['default']
     MailLogo: typeof import('./src/components/Icons/MailLogo.vue')['default']
     MailThread: typeof import('./src/components/MailThread.vue')['default']
     MailThreadPlaceholder: typeof import('./src/components/MailThreadPlaceholder.vue')['default']
@@ -48,7 +49,6 @@ declare module 'vue' {
     SendMail: typeof import('./src/components/SendMail.vue')['default']
     SendMailMobileLayout: typeof import('./src/components/SendMailMobileLayout.vue')['default']
     SettingsModal: typeof import('./src/components/Modals/SettingsModal.vue')['default']
-    SidebarDetail: typeof import('./src/components/SidebarDetail.vue')['default']
     SidebarLink: typeof import('./src/components/SidebarLink.vue')['default']
     UserDropdown: typeof import('./src/components/UserDropdown.vue')['default']
   }
